# Conflicts:
#	src/Controleur/MoteurJeu.java
#	src/Global/Config.java
#	src/IHM/Graphique/Ecrans/EcranCreationPartie.java
#	src/IHM/Graphique/IHMGraphique.java
#	src/TestJeu.java

+++ src/Modele/IA/Heuristique.java
@@ -1,7 +1,7 @@
 package Modele.IA;
 
-import Modele.Jeu.Jeu;
+import Modele.Jeux.Jeu;
 
 public interface Heuristique {
-    public int evaluer(Jeu j, int pdvJoueur);
+    public double evaluer(Jeu j, int pdvJoueur);
 }
